Fixed crash when using energymeter mod without computercraft
