+++ src/main/java/com/almostreliable/energymeter/EnergyMeter.java
@@ -2,6 +2,7 @@ package com.almostreliable.energymeter;
 
 import com.almostreliable.energymeter.core.Config;
 import com.almostreliable.energymeter.core.Registration;
+import com.almostreliable.energymeter.compat.CapabilityAdapterFactory;
 import com.almostreliable.energymeter.data.DataGeneration;
 import com.almostreliable.energymeter.network.PacketHandler;
 
@@ -22,6 +23,7 @@ public final class EnergyMeter {
 
     public EnergyMeter(IEventBus modEventBus, ModContainer modContainer) {
         Registration.init(modEventBus);
+        CapabilityAdapterFactory.init(modEventBus);
         PacketHandler.init(modEventBus);
         Config.init(modContainer);
         modEventBus.addListener(DataGeneration::init);
